Unreviewed, reverting 276672@main. https://bugs.webkit.org/show_bug.cgi?id=271695
Broke with newer SDK

Reverted changeset:

"Use std::span more with SharedBuffer"
https://bugs.webkit.org/show_bug.cgi?id=271534
https://commits.webkit.org/276672@main

Canonical link: https://commits.webkit.org/276676@main

+++ Source/WTF/wtf/FileSystem.h
@@ -168,9 +168,9 @@ WTF_EXPORT_PRIVATE long long seekFile(PlatformFileHandle, long long offset, File
 WTF_EXPORT_PRIVATE bool truncateFile(PlatformFileHandle, long long offset);
 WTF_EXPORT_PRIVATE bool flushFile(PlatformFileHandle);
 // Returns number of bytes actually read if successful, -1 otherwise.
-WTF_EXPORT_PRIVATE int64_t writeToFile(PlatformFileHandle, const void* data, size_t length);
+WTF_EXPORT_PRIVATE int writeToFile(PlatformFileHandle, const void* data, int length);
 // Returns number of bytes actually written if successful, -1 otherwise.
-WTF_EXPORT_PRIVATE int64_t readFromFile(PlatformFileHandle, void* data, size_t length);
+WTF_EXPORT_PRIVATE int readFromFile(PlatformFileHandle, void* data, int length);
 
 WTF_EXPORT_PRIVATE PlatformFileHandle openAndLockFile(const String&, FileOpenMode, OptionSet<FileLockMode> = FileLockMode::Exclusive);
 WTF_EXPORT_PRIVATE void unlockAndCloseFile(PlatformFileHandle);
